web/requirements.txt
Django==2.2
djangorestframework==3.10.1
psycopg2-binary==2.8.3
Pillow
requests
django-debug-toolbar
git+https://github.com/mirsoft-kg/sms-sender.git
python-slugify
